btle: ammended clock source selection
In btl/btle.cpp there was a forced dependancy on a Targets'
macro used during compilation. The more targets that are
added, the longer this macro would get. I have refined
the macro to look for an RC source which is specified during
compile time.

This will also prevent any serious headaches for anyone in the
future who wonders at 3am why BLE crashes the device.