microbit: Reduced default GATT table size
Soft Device holds a dedicated GATT table at the top of SD allocated memory.
This is of configurable size as of S110 8.0, so we reduce this to suit the minimum set
of BLE service for the micro:bit. The remaining memory can then be recycle as heap
memory using the microbit heap allocator.